+++ CHANGES.md
@@ -1,11 +1,13 @@
 wavesurfer.js changelog
 =======================
+
 6.0.0 (unreleased)
 ------------------
 - Add additional type to `waveColor` and `progressColor` parameters to support linear gradients (#2345)
 - Add `hideCursor` option to hide the mouse cursor when hovering over the waveform (#2367)
 - Add optional `channelIdx` parameter to `setWaveColor`, `getWaveColor`, `setProgressColor` and
   `getProgressColor` methods (#2391)
+- Improved drawing waveform with bars, now bars height is the maximum peak value in range (#2428) 
 - Markers plugin: Add the ability to set markers as draggable using param `draggable=true`,
   `marker-drag` and `marker-drop` events will be triggered (#2398)
 - Regions plugin: